backend/requirements.txt
# Flask and extensions
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Caching==2.0.2
PyJWT==2.6.0
Werkzeug==2.2.3

# Data processing
pandas==1.5.3
numpy==1.24.2
yfinance==0.2.12
requests==2.28.2

# Google Cloud
google-cloud-storage==2.7.0

# Utilities
python-dotenv==1.0.0
gunicorn==20.1.0